LayoutTests/fast/dom/getElementsByTagName-HTMLElement-prefix.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<wx:map id="testElement"></wx:map>
<script>
description("Tests that getElementsByTagName() works properly for elements that have a semicolon in their tag name, inside an HTML document.");

var testElement = document.getElementById("testElement");
shouldBeNull("testElement.prefix");
shouldBeEqualToString("testElement.localName", "wx:map");
shouldBeEqualToString("testElement.tagName", "WX:MAP");

evalAndLog('collection = document.getElementsByTagName("wx:map")');
shouldBe("collection.length", "1");
shouldBe("collection[0]", "testElement");

evalAndLog('collection = document.getElementsByTagName("WX:MAP")');
shouldBe("collection.length", "1");
shouldBe("collection[0]", "testElement");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
